pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-composer"
version = "0.1.0"
description = "AIComposer — a tool for generating verified implementations from documentation and CVL specifications"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Certora Ltd."}
]

dependencies = [
    "graphcore>=0.1.0",
    "aiohttp>=3.13",
    "attrs>=26.1",
    "Jinja2>=3.1",
    "langgraph-checkpoint-postgres>=3.0",
    "numpy>=2.4",
    "prompt-toolkit>=3.0",
    "psycopg[binary]>=3.2",
    "pydantic>=2.12",
    "questionary>=2.1",
    "rich>=14.3",
    "sentence-transformers>=5.1",
    "textual>=8.1",
    "websockets",
    "chromadb>=0.4.0",
    "aiosqlite"
]

[dependency-groups]
test = ["pytest>=9.0", "pytest-asyncio>=1.3", "testcontainers>=4.0"]
ci = ["pyright"]

[tool.uv.sources]
graphcore = { path = "./graphcore", editable = true }

[project.scripts]
cex-analyzer = "analyzer.analysis:main"

[tool.setuptools.packages.find]
include = ["composer*"]

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.package-data]
composer = ["templates/*.j2"]